refactor(mpt): remove unused utils.rs
